clean up and add comment to configuration object
